+++ SenseCareLocal/Models/VitalSigns/VitalSign.cs
@@ -16,7 +16,7 @@ public class VitalSign
     public bool Fuente { get; set; }
 
     [BsonElement("pulso")]
-    public List<int> Pulso { get; set; } = new ();
+    public List<double> Pulso { get; set; } = new ();
 
     [BsonElement("temperatura")]
     public double? Temperatura { get; set; }
